feat: Basic development setup
Add a basic Compose file, which makes it possible to run the core of TMS
in containers. Optional features are not supported yet. (Some, like the
evaluator or the JPlag plagiarism check, should be refactored out from
the backend core and use separate containers. Others, like version-
controlled submissions, may have place in a backend core with more APT
packages installed.) There is a frontend container, both the backend and
frontend containers load the code from bind mounts, and phpMyAdmin is
configured to automatically log the user in – all of these make this
setup unsuitable for production use. (A production setup should use only
one container, containing both the backend code and the compiled
frontend code, as part of the image rather than in mounted directories.)